Lots of changes.
I found a few wiring mistakes in the schematic and fixed them up.
I added microcode instructions to allow B to be copied out to A,
to the UART and to memory. I added a lot of `mem = mem op B`
instructions. I added some examples to test the above. I redid the
PCB layout and added copper pours.